add delete_indexes to index_management
add delete indexes by index id patterns
